pyproject.toml
[tool.poetry]
name = "video-crop"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <murumuru1204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
moviepy = "^1.0.3"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
pysen = {version = "0.10.1", extras = ["lint"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py39"

[tool.pysen.lint.source]
excludes = [".venv/"]

[[tool.pysen.lint.mypy_targets]]
paths = ["."]
